Addressed code styling issues, and performed file reorganization
