+++ iACC/DispatchQueue+MainThread.swift
@@ -1,0 +1,15 @@
+//	
+// Copyright © Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+extension DispatchQueue {
+    static func mainAsyncIfNeeded(execute work: @escaping () -> Void) {
+        if Thread.isMainThread {
+            work()
+        } else {
+            main.async(execute: work)
+        }
+    }
+}
