+++ iACC/Formatters.swift
@@ -1,0 +1,10 @@
+//	
+// Copyright © Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+enum Formatters {
+    static var date = DateFormatter()
+    static var number = NumberFormatter()
+}
