+++ iACC/UIViewController+Presentation.swift
@@ -4,11 +4,6 @@
 
 import UIKit
 
-enum Formatters {
-    static var date = DateFormatter()
-    static var number = NumberFormatter()
-}
-
 extension UIViewController {
     var presenterVC: UIViewController {
         parent?.presenterVC ?? parent ?? self
@@ -23,12 +18,14 @@ extension UIViewController {
     func select(card: Card) {
         let vc = CardDetailsViewController()
         vc.card = card
-        show(vc, sender: self)    }
+        show(vc, sender: self)
+    }
     
     func select(transfer: Transfer) {
         let vc = TransferDetailsViewController()
         vc.transfer = transfer
-        show(vc, sender: self)    }
+        show(vc, sender: self)
+    }
     
     @objc func addCard() {
         show(AddCardViewController(), sender: self)
@@ -51,15 +48,4 @@ extension UIViewController {
         alert.addAction(UIAlertAction(title: "Ok", style: .default))
         showDetailViewController(alert, sender: self)
     }
-}
-
-
-extension DispatchQueue {
-	static func mainAsyncIfNeeded(execute work: @escaping () -> Void) {
-		if Thread.isMainThread {
-			work()
-		} else {
-			main.async(execute: work)
-		}
-	}
 }
